==> notebooks/01_Data_Ingestion_and_Cleaning.ipynb <==
### Initialize

# Import libraries
import pandas as pd
import numpy as np
import os

# Configure path to data folder
DATA_PATH = 'C:/Users/USER/Documents/Portfolio stuff/Fintech_CLV_Churn_Analysis/data/'
SAMPLE_FRACTION = 0.10 # 10% sample for the largest file (order_products__prior)

# Load datasets
orders = pd.read_csv(os.path.join(DATA_PATH, 'orders.csv'))
products = pd.read_csv(os.path.join(DATA_PATH, 'products.csv'))
departments = pd.read_csv(os.path.join(DATA_PATH, 'departments.csv'))

num_lines = sum(1 for line in open(os.path.join(DATA_PATH, 'order_products__prior.csv'))) - 1 # Calculate the number of lines (required for sampling logic)
skip_idx = np.random.choice(np.arange(1, num_lines + 1), size=int(num_lines * (1 - SAMPLE_FRACTION)), replace=False) # Create an index of rows to skip (we skip 90 out of every 100 rows, plus the header)

# Load the sampled data
order_products_prior = pd.read_csv(os.path.join(DATA_PATH, 'order_products__prior.csv'), skiprows=skip_idx)

### Merging the files to create a transaction log (since this is framing the analysis into a fintech frame)

# Merge order line items with product details to get department info.
# Join 'order_products_prior' (the transaction details) with 'products' to associate each line item with its department (the proxy for 'Merchant Vertical').
merged_df = order_products_prior.merge(products[['product_id', 'department_id']], on='product_id', how='left')

# Merge the combined product details with the order metadata (customer ID, time) to create the actual transaction log.
# join with 'orders' to link each transaction (product line item) back to the specific 'user_id' (Customer ID) and the time.
transaction_log = merged_df.merge(orders[['order_id', 'user_id', 'order_number', 'days_since_prior_order', 'order_dow', 'order_hour_of_day', 'eval_set']], on='order_id', how='left')

### Little EDA

print(f"\nInitial Transaction Log Merged. \nShape: {transaction_log.shape}")

print(transaction_log.info())
transaction_log.head()

transaction_log.describe(include='all')

transaction_log.isnull().sum()

# View dataframe of only null days_simce_prior_order
transaction_log[transaction_log['days_since_prior_order'].isnull()]

# convert days_since_prior_order to int
transaction_log['days_since_prior_order'] = transaction_log['days_since_prior_order'].fillna(0).astype(int)

this exploration shows that NaN values in days_since_prior_orders indeed reflect every customer's first order

# filtering out test/train sets
print(transaction_log[transaction_log['eval_set'] == 'train'])
transaction_log[transaction_log['eval_set'] == 'test']

somehow, every test/train eval_set was not selected during the sampling. hm

# drop eval_set column as it's not needed for analysis
transaction_log.drop(columns=['eval_set'], inplace=True)

# checking for duplicates
transaction_log.duplicated().sum()

# Renaming key columns to match Fintech Framing
transaction_log.rename(columns={
    'user_id': 'Customer_ID',
    'order_id': 'Transaction_ID',
    'department_id': 'Merchant_Vertical_ID',
    'product_id': 'Product_SKU' 
}, inplace=True)


# Select and order final columns for clarity
final_transaction_log = transaction_log[['Customer_ID', 'Transaction_ID', 'order_number', 'days_since_prior_order', 'order_dow', 'order_hour_of_day', 'Merchant_Vertical_ID', 'Product_SKU']]

final_transaction_log.head()

# Save the final, clean, merged log to the data folder
OUTPUT_FILE = os.path.join(DATA_PATH, 'processed_features.csv')
final_transaction_log.to_csv(OUTPUT_FILE, index=False)

print(f"Final shape of transaction log: {final_transaction_log.shape}")
